+++ api/routes/index.py
@@ -1,5 +1,5 @@
 from __main__ import app
-from flask import render_template, jsonify, redirect, url_for
+from flask import render_template, redirect, url_for
 import requests
 
 from func.token import get_access_token
@@ -8,13 +8,14 @@ from func.header import get_auth_header
 # import other routes here
 from routes.test import test
 from routes.auth.index import *
+from routes.search.routes import  searchArtist, searchTrack
 
 
 @app.route('/api')
 def index():
     token = get_access_token()
     if not token:
-        return redirect(url_for('auth.index'))
+        return render_template("404.html")
         
     # Verify token by making a test request to Spotify API
     headers = get_auth_header(token)
@@ -23,4 +24,4 @@ def index():
         response.raise_for_status()
         return render_template('index.html')
     except requests.exceptions.RequestException:
-        return redirect(url_for('auth.index'))
+        return redirect(url_for('404.html'))