+++ next.config.ts
@@ -31,6 +31,16 @@ const nextConfig: NextConfig = {
             },
         ];
     },
+    images: {
+        remotePatterns: [
+            {
+                protocol: "https",
+                hostname: "i.scdn.co",
+                port: "",
+                pathname: "/image/**",
+            },
+        ]
+    },
 }
 
 export default nextConfig;