+++ src/components/Header.tsx
@@ -1,13 +1,13 @@
-import Image from "next/image"
 import Link from "next/link"
+import { ImageComponent } from "./ui/ImageComponent"
 
 export const Header = () => {
     return (
         <header style={{
             display: "flex",
-            justifyContent: "center",
+            justifyContent: "start",
             alignItems: "center",
-            padding: "5rem 2.5rem",
+            padding: "6rem 2.5rem",
             gap: "1rem",
         }}>
             <h1>
@@ -15,7 +15,14 @@ export const Header = () => {
                     Spotify API Starter
                 </Link>
             </h1>
-            <Image src="/android-chrome-512x512.png" alt="Spotify API Starter" width={50} height={50} />
+            <ImageComponent 
+                image={{
+                    src: "/android-chrome-512x512.png",
+                    alt: "Spotify API Starter",
+                    width: 64,
+                    height: 64,
+                }} 
+            />
         </header>
     )
 }