+++ tailwind.config.js
@@ -5,6 +5,9 @@ module.exports = {
     './src/**/*.{ts,tsx,js,jsx}',
   ],
   theme: {
-    extend: {},
+    extend: {
+    
+    },
   },
+  
 };