+++ CHANGELOG.md
@@ -1,14 +1,21 @@
 # Changelog
-> Last updated: `2025-02-08`
-> - [Accessed user data on the frontend](#accessed-user-data-on-the-frontend)
->   - [@clxrityy](https://github.com/clxrityy)
+> Last updated: `2025-02-12`
+> - [Search for artists and tracks](#search-for-artists-and-tracks)
 
 # OUTLINE
 
-- [Initial commit & setup](#initial-commit--setup) | [@clxrityy](https://github.com/clxrityy)
-- [Setting up the frontend (Next.js)](#setting-up-the-frontend-nextjs) | [@clxrityy](https://github.com/clxrityy)
-- [Implement Authorization](#implement-authorization) | [@clxrityy](https://github.com/clxrityy)
-- [Accessed user data on the frontend](#accessed-user-data-on-the-frontend) | [@clxrityy](https://github.om/clxrityy)
+- ```2025-02-06``` | [Initial commit & setup](#initial-commit--setup)
+- ```2025-02-07``` | [Setting up the frontend (Next.js)](#setting-up-the-frontend-nextjs)
+    - [NPM Setup](#npm-setup)
+    - [App structure setup](#app-structure-setup)
+    - [Integrating the API with the frontend](#integrating-the-api-with-the-frontend)
+-  ```2025-02-08``` | [Implement Authorization](#implement-authorization)
+- ```2025-02-08``` | [Accessed user data on the frontend](#accessed-user-data-on-the-frontend)
+    - [Authorization flow](#authorization-flow)
+    - [Accessing user data](#accessing-user-data)
+- ```2025-02-10``` | [Top 10 Artists & Tracks](#top-10-artists--tracks)
+- ```2025-02-12``` | [Search for artists and tracks](#search-for-artists-and-tracks)
+    - [Client-side display](#client-side-display)
 
 ## Contributers
 
@@ -22,7 +29,6 @@ Whenever you make a change to the project, please update the changelog with a br
     # Changelog
     > Last updated: `2025-02-06`
     > - [Brief title](#brief-title)
-    >   - [@username](https://github.com/username)
     ```
 2. Add your change
     - Create a new heading for your change in the format `## Brief title`.
@@ -37,7 +43,7 @@ Whenever you make a change to the project, please update the changelog with a br
     - Include any link(s) or external resources that are relevant to the change.
     - Add a link to the heading to the changelog outline (in its respective section).
     ```md
-    - [Brief title](#brief-title) | [@username](https://github.com/username)
+    - ```YYYY-MM-DD``` [Brief title](#brief-title)
     ```
     - If applicable, add a `##### Resources` and/or `##### References` section with links to relevant resources.
 3. Add your changelog entry to the outline at the top of the page.
@@ -205,3 +211,66 @@ Whenever you make a change to the project, please update the changelog with a br
         - Reference: [Setting cookie in server action gives me an error | Reddit](https://www.reddit.com/r/nextjs/comments/1flih2p/comment/lo6e83l/?utm_source=share&utm_medium=web3x&utm_name=web3xcss&utm_term=1&utm_content=share_button)
 - [Authorization Code Flow with Proof Key for Code Exchange (PKCE) | Spotify for Developers](https://developer.spotify.com/documentation/web-api/tutorials/code-pkce-flow)
 - [Error: NEXT_REDIRECT](https://stackoverflow.com/questions/76191324/next-13-4-error-next-redirect-in-api-routes)
+
+---
+
+## Top 10 Artists & Tracks
+
+> `2025-02-10` | [@clxrityy](https://github.com/clxrityy)
+
+When clicking the user button (top right corner) (must be logged in), you can see your top 10 artists and tracks.
+- The data is fetched from the backend API endpoint (`/api/auth/profile/top/<type>`) where `<type>` can be either `artists` or `tracks`.
+    - The endpoint fetches the user's top 10 artists or tracks from Spotify.
+- The data is displayed in a modal on the frontend.
+
+### References
+
+- [`app/profile/[id]/page.tsx`](/src/app/(routes)/profile/[id]/page.tsx)
+- [`api/routes/auth/profile/top.py`](/api//routes/auth/profile/top.py)
+
+---
+
+## Search for artists and tracks
+
+> `2025-02-12` | [@clxrityy](https://github.com/clxrityyy)
+
+You can now search for artists and tracks using the search bar(s) on the homepage.
+- The search bar queries the backend API endpoint (`/api/search/<type>/<query>`) where `<type>` can be either `artist` or `track`.
+    - The endpoint fetches the search results from Spotify. (`https://api.spotify.com/v1/search`)
+
+### Client-side display
+
+- The search results are displayed as the query is typed.
+    - If the user clicks out of the area of focus, the search results are no longer visible.
+- (For now) only **1** artist or track is displayed at a time.
+
+Many hooks and components were created to handle the search functionality.
+
+- Installed `lodash` for debouncing the search queries.
+
+```tsx
+"use client";
+// ...
+import {
+    useCallback,
+    // ...
+} from "react";
+import debouce from "lodash/debounce";
+
+const debouncedSearch = useCallback(debouce(async (searchQuery: string) => {
+    // ...
+}, 300), [
+    // ...
+])
+```
+
+> See [`Search.tsx`](/src/components/ui/Search.tsx) for more details.
+
+### References
+
+- [`api/routes/search/routes.py`](/api/routes/search/routes.py)
+- [`src/components/ui/Search.tsx`](/src/components/ui/Search.tsx)
+- [`src/components/SearchResults.tsx`](/src/components/SearchResults.tsx)
+- [Web API Reference | Search](https://developer.spotify.com/documentation/web-api/reference/search)
+- [Debouncing in React | Medium](https://medium.com/@manigandham/debouncing-react-events-with-react-hooks-2f1d6f5c2467)
+- [How to use debounce hooks in React](https://stackoverflow.com/questions/75556418/how-to-use-debounce-hooks-in-react)